moved helper functions from backend to utils
